Add middleware for setting cookie message cookie
This middleware responds to a POST request with the params
`cookies=accepted` by setting the `euCookieNotice` cookie
to `accepted` on the request's host with a path of `/`. It
is also set to expire in one year.

The middleware then redirects to the referrer if one is
specified or the original request url if one isn't. The
referrer must be on the same host and port otherwise it
falls back to the original request url. This is to prevent
use for open redirection[1].

[1]: https://webmasters.googleblog.com/2009/01/open-redirect-urls-is-your-site-being.html