Reduce logging by cookie message middleware
The syslogs.nature.com server is being spammed by logs from the cookie
message middleware because it logs at the default :info level for
production apps. To reduce the volume of output, log at the debug level
and provide a way to override the logger if more info is needed in
a production application.